+++ dto/forecast_result.go
@@ -1,7 +1,7 @@
 package dto
 
 type ForecastResult struct {
-    Timestamp   string  `json:"timestamp"`
-    Temperature float64 `json:"temperature"`
-    PrecipitationProbability int     `json:"precipitationProbability"`
+    Timestamp                string  `json:"timestamp"`
+    Temperature              float64 `json:"temperature"`
+    PrecipitationProbability float64 `json:"precipitationProbability"`
 }
